fix: multi-source 4326→3857 resampling to eliminate cliff walls at tile boundaries
When a 3857 tile straddles two 4326 tiles, the old single-source
resampling clamped edge heights, creating visible cliff walls.

New approach:
- Cache rasterized 4326 heightmaps by (gx, gy, gz)
- Detect overlapping 4326 tiles via overlapping_geographic_tiles()
- resample_geographic_to_mercator_multi() picks the correct source per-pixel
- Re-resample previously loaded adjacent tiles when new 4326 data arrives
- Invalidate mesh cache for re-resampled tiles

Also: refactor NativeApp into modules (app/mod, init, input, render_layers,
tile_loading, tile_upload, tiles3d_frame), add animation easing, TileJSON
resolution, 3D Tiles traversal improvements, and various fixes.

+++ crates/x-planets-core/src/viewport.rs
@@ -90,7 +90,7 @@ impl Viewport {
             cam_h * pitch_rad.cos(),
         );
         let target = glam::DVec3::new(cx, cy, 0.0);
-        let up_hint = glam::DVec3::new(sin_b as f64, -cos_b as f64, 0.0);
+        let up_hint = glam::DVec3::new(sin_b, -cos_b, 0.0);
 
         // Camera basis (matches glam look_at_rh convention).
         let f = (target - eye).normalize();
